have the new settings menu handle its own state
